requirements.txt
# Web UI
streamlit>=1.32.0

# HTTP & scraping
requests>=2.31.0
curl_cffi>=0.14.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data
pandas>=2.2.0
python-dateutil>=2.9.0

# Scheduling
apscheduler>=3.10.4

# Feed parsing (kept for potential future use)
feedparser>=6.0.11

# Document parsing (resume upload)
pypdf>=3.0.0
python-docx>=1.2.0

# AI providers
groq>=1.0.0
google-generativeai>=0.8.0
